improve github module
