+++ src/github/internal/fetch.ts
@@ -1,16 +1,18 @@
 import { ENV_CONFIG } from "../../env-config.js";
 import { request } from "@octokit/request";
+import { OctokitResponse } from "@octokit/types";
+import { GithubFetchError } from "../domain.js";
 
-export const fetchWithAuth = request.defaults({
+export const fetchFromGithub = request.defaults({
   headers: {
     authorization: `token ${ENV_CONFIG.GITHUB_TOKEN}`,
   },
 });
 
-export type GithubResponse = Awaited<ReturnType<typeof request>>;
-
-export class GithubFetchError extends Error {
-  constructor(response: GithubResponse) {
-    super(`Failed to fetch ${response.url}, received status ${response.status}`);
+export function getResponseData<T>(response: OctokitResponse<T>) {
+  if (response.status !== 200) {
+    throw new GithubFetchError(response);
   }
-}
+
+  return response.data;
+};
